remove boilerplate from test file
